+++ src/index.ts
@@ -1,9 +1,5 @@
-/*
- * Public API Surface of ngxf-uploader
- */
 export * from './lib/ngxf-drop.directive';
 export * from './lib/ngxf-select.directive';
 export * from './lib/ngxf-parse.directive';
 export * from './lib/ngxf-uploader.service';
-export * from './lib/ngxf-uploader.module';
 export * from './lib/ngxf-uploader.model';