+++ src/lib/ngxf-drop.directive.ts
@@ -1,20 +1,21 @@
 import {
   Directive,
   ElementRef,
-  EventEmitter,
   HostListener,
-  Input,
-  Output,
   Renderer2,
+  inject,
 } from '@angular/core';
 
 import {
   FileError,
-  FileOption,
+  FileValidateOptions,
   NgxfDirectoryStructure,
   NgxfUploadDirective,
+  NgxfUploadFolder,
 } from './ngxf-uploader.model';
 import { getFilesFromGetAsEntry, getUploadResult } from './utils';
+import { input } from '@angular/core';
+import { output } from '@angular/core';
 
 const stopEvent = (e: Event) => {
   e.stopPropagation();
@@ -26,44 +27,47 @@ const stopEvent = (e: Event) => {
  */
 @Directive({
   selector: '[ngxf-drop]',
+  standalone: true,
 })
 export class NgxfDropDirective implements NgxfUploadDirective {
+  private _elm = inject(ElementRef);
+  private _render = inject(Renderer2);
   /** when get drop file into area that will trigger */
-  @Output('ngxf-drop') onUpload = new EventEmitter<
-    File | File[] | NgxfDirectoryStructure[] | FileError
-  >();
+  onUpload = output<File | File[] | NgxfDirectoryStructure[] | FileError>({
+    alias: 'ngxf-drop',
+  });
 
-  @Input('ngxf-validate') fileOption: FileOption = {};
-  @Input() multiple!: string;
-  @Input() accept!: string;
-  @Input() structure!: NgxfUploadDirective['structure'];
+  fileOption = input<FileValidateOptions>({}, { alias: 'ngxf-validate' });
+  multiple = input<boolean | undefined>();
+  accept = input<string | undefined>();
+  structure = input<NgxfUploadFolder>();
 
   /**
    * add class when drop into this drop area
    * @default 'drop'
    */
-  @Input('drop-class') dropClass = 'drop';
-
-  constructor(private _elm: ElementRef, private _render: Renderer2) {}
+  dropClass = input('drop', { alias: 'drop-class' });
 
   @HostListener('drop', ['$event']) async drop(e: DragEvent) {
     stopEvent(e);
 
-    this._render.removeClass(this._elm.nativeElement, this.dropClass);
+    this._render.removeClass(this._elm.nativeElement, this.dropClass());
 
-    let files: FileList | File[];
+    let files: FileList | File[] | undefined;
     try {
       files = await getFilesFromGetAsEntry(e);
     } catch (error) {
-      files = e.dataTransfer.files;
+      files = e.dataTransfer?.files;
     }
+
+    if (!files) return;
 
     const result = getUploadResult(
       files,
-      this.accept,
-      this.multiple,
-      this.fileOption,
-      this.structure
+      this.accept() || '',
+      this.multiple(),
+      this.fileOption(),
+      this.structure(),
     );
 
     this.onUpload.emit(result);
@@ -73,11 +77,11 @@ export class NgxfDropDirective implements NgxfUploadDirective {
   @HostListener('dragenter', ['$event'])
   dragenter(e: Event) {
     stopEvent(e);
-    this._render.addClass(this._elm.nativeElement, this.dropClass);
+    this._render.addClass(this._elm.nativeElement, this.dropClass());
   }
 
   @HostListener('dragleave', ['$event']) dragleave(e: Event) {
     stopEvent(e);
-    this._render.removeClass(this._elm.nativeElement, this.dropClass);
+    this._render.removeClass(this._elm.nativeElement, this.dropClass());
   }
 }