+++ src/lib/utils/upload-error.ts
@@ -1,0 +1,10 @@
+import { UploadEvent } from '../ngxf-uploader.model';
+
+export class NgxfUploadError extends Error {
+  constructor(
+    message: string,
+    private state: UploadEvent,
+  ) {
+    super(message);
+  }
+}